+++ src/main/java/uz/jl/models/card/Card.java
@@ -6,6 +6,7 @@ import uz.jl.enums.card.CardType;
 import uz.jl.models.Auditable;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  * @author Elmurodov Javohir, Mon 12:07 PM. 12/6/2021
@@ -22,7 +23,7 @@ public class Card extends Auditable {
     private String password;
     private CardType type;
     private Status status;
-    private BigDecimal balance;
+    private BigInteger balance;
     private String bankId;
     private String holderId;
     private String smsPhoneNumber;
